goood

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -8,7 +8,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	unsigned int count = 0;
-	unsigned int long p = (n ^ m);
+	unsigned int p = (n ^ m);
 
 	while (p)
 	{
